[ClrPhLib] Update ClrPhLib now that phlib is in third_party folder

+++ ClrPhlib/src/managed/PeImport.cpp
@@ -91,7 +91,7 @@ PeImportDll::PeImportDll(
 
 	for (size_t IndexImport = 0; IndexImport < (size_t)NumberOfEntries; IndexImport++)
 	{
-		ImportList->Add(gcnew PeImport(other->ImportList[IndexImport]));
+		ImportList->Add(gcnew PeImport(other->ImportList[(int) IndexImport]));
 	}
 
 }
